# Conflicts:
#	doc/admin/executors/deploy_executors_kubernetes.md
#	enterprise/cmd/batcheshelper/run/pre.go
#	enterprise/cmd/executor/internal/config/config.go
#	enterprise/cmd/executor/internal/config/config_test.go
#	enterprise/cmd/executor/internal/run/util.go
#	enterprise/cmd/executor/internal/worker/command/kubernetes.go
#	enterprise/cmd/executor/internal/worker/runner/kubernetes.go
#	enterprise/cmd/executor/kubernetes/executor-batches.Deployment.yml

+++ enterprise/cmd/executor/internal/worker/runner/kubernetes.go
@@ -56,13 +56,15 @@ func (r *kubernetesRunner) TempDir() string {
 }
 
 func (r *kubernetesRunner) Teardown(ctx context.Context) error {
-	for _, name := range r.jobNames {
-		if err := r.cmd.DeleteJob(ctx, r.options.Namespace, name); err != nil {
-			r.internalLogger.Error(
-				"Failed to delete kubernetes job",
-				log.String("jobName", name),
-				log.Error(err),
-			)
+	if !r.options.KeepJobs {
+		for _, name := range r.jobNames {
+			if err := r.cmd.DeleteJob(ctx, r.options.Namespace, name); err != nil {
+				r.internalLogger.Error(
+					"Failed to delete kubernetes job",
+					log.String("jobName", name),
+					log.Error(err),
+				)
+			}
 		}
 	}
